chore(update-verifier): use correct Cli::parse
